+++ app/src/main/java/com/example/esieaboard/EventPageActivity.java
@@ -1,11 +1,13 @@
 package com.example.esieaboard;
 
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.database.Cursor;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.TextView;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import com.example.esieaboard.admin.EditEventActivity;
@@ -23,7 +25,7 @@ public class EventPageActivity extends AppCompatActivity {
     private static final int EDIT_EVENT_REQUEST_CODE = 1;
 
     ImageButton backButton;
-    Button editButton, attendButton, cancelButton;
+    Button editButton, attendButton, cancelButton, deleteButton;
     TextView eventName, eventDate, eventLocation, eventDescription, eventCapacity, eventAttendees;
     AttendanceModel attendance;
     EventModel event;
@@ -41,6 +43,7 @@ public class EventPageActivity extends AppCompatActivity {
         editButton = findViewById(R.id.edit_button);
         attendButton = findViewById(R.id.attend_button);
         cancelButton = findViewById(R.id.cancel_button);
+        deleteButton = findViewById(R.id.delete_button);
         eventName = findViewById(R.id.event_name);
         eventDate = findViewById(R.id.event_date);
         eventLocation = findViewById(R.id.event_location);
@@ -65,9 +68,17 @@ public class EventPageActivity extends AppCompatActivity {
             attendButton.setVisibility(GONE);
         }
 
+        deleteButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                confirmDialog();
+            }
+        });
+
         backButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
+                setResult(RESULT_OK);
                 finish();
             }
         });
@@ -90,6 +101,7 @@ public class EventPageActivity extends AppCompatActivity {
                 dataBaseHelper.addAttendance(attendance);
                 attendButton.setVisibility(GONE);
                 cancelButton.setVisibility(VISIBLE);
+                eventAttendees.setText(String.valueOf(countAttendance()));
             }
         });
 
@@ -108,6 +120,7 @@ public class EventPageActivity extends AppCompatActivity {
                 dataBaseHelper.deleteAttendanceRow(String.valueOf(attendance.getId()));
                 attendButton.setVisibility(VISIBLE);
                 cancelButton.setVisibility(GONE);
+                eventAttendees.setText(String.valueOf(countAttendance()));
             }
         });
     }
@@ -169,7 +182,28 @@ public class EventPageActivity extends AppCompatActivity {
         }
     }
 
-    // In UserProfileActivity.java
+    void confirmDialog() {
+        AlertDialog.Builder builder = new AlertDialog.Builder(EventPageActivity.this);
+        builder.setTitle("Delete " + event.getName() + " ?");
+        builder.setMessage("Do you really want to delete " + event.getName() + " ?");
+        builder.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialogInterface, int i) {
+                DataBaseHelper dataBase = new DataBaseHelper(EventPageActivity.this);
+                dataBase.deleteEventRow(String.valueOf(event.getId()));
+                setResult(RESULT_OK);
+                finish();
+            }
+        });
+        builder.setNegativeButton("No", new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialogInterface, int i) {
+
+            }
+        });
+        builder.create().show();
+    }
+
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);