+++ app/src/main/java/com/example/esieaboard/SignInActivity.java
@@ -4,6 +4,7 @@ import android.provider.ContactsContract;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import com.example.esieaboard.models.UserModel;
@@ -28,10 +29,18 @@ public class SignInActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 DataBaseHelper dataBase = new DataBaseHelper(SignInActivity.this);
-                UserModel user = new UserModel(-1, "", "",
-                        inputEmail.getText().toString(), inputPassword.getText().toString(), "", 0);
-                dataBase.addUser(user);
-                finish();
+                if(dataBase.getUserByEmail(inputEmail.getText().toString().trim()) == null)
+                {
+                    UserModel user = new UserModel(-1, "User", "",
+                            inputEmail.getText().toString().trim(), inputPassword.getText().toString().trim(), "", 0);
+                    dataBase.addUser(user);
+                    finish();
+                }
+                else {
+                    Toast.makeText(SignInActivity.this, "This user already exists", Toast.LENGTH_SHORT).show();
+                    inputEmail.setText("");
+                    inputPassword.setText("");
+                }
             }
         });